+++ src/lang/parser.hpp
@@ -11,6 +11,7 @@
 #include <sstream>
 #include <filesystem>
 #include <map>
+#include <iostream>
 
 namespace cones {